Add ENV config options for uWSGI limits
It is handy to be able to set uWSGI parameters via ENV in a container
context, especially because those limits are highly dependant on the
use-case so setting them is a reasonably common case.

I've added the listen configuration option with a default of 256. The
uWSGI default is 100 I think, which is quite low for setups with more
serious load. A modern kernel / OS usually is configured for 1024
concurrent connections (aka open file descriptors) so 256 should not be
too bad. In production I actually use 512 to cope with the load,
otherwise I get 502 resource unavailable errors, but that just as a
side-note.

I've kept the limits in the uWSGI config file and added the listen
option there too in case a setup uses a custom supervisord.conf but the
container packaged uwsgi.ini.