Make supervisord log level configurable via env
The supervisord log level controls the containers log output regardless
of the DEBUG=0|1 flag or other nginx log level settings, thus should be
configurable.

IMO I'd set log level to warn by default to supress access logs which
is desired for most production deployments I suppose. That really depends
on the main use-case for this image though. If it's for development setups,
then of course debug is totally fine. Also debug won't brake existing
behavior.

DOCS: the docs need an update to indicate that this can be controlled
for docker containers. While we are at it I'd also add the other options
of hipervisord that can be controlled this way.